🐛 fix: 修复 star 追踪 release 事件归属错误与中断补拉丢失，subject_number 升级 int64
- 修复 release 事件归属张冠李戴：外部 star 仓不建 Repository 行，release 事件落库补写 PayloadSummary.repository，摘要预览与 AI 输入统一经 store.EventRepoName 解析仓库名（RepositoryID → 回退 PayloadSummary），star/watch 事件标题即仓库名时不再重复前缀
- 修复 release 轮询中断后中间版本静默丢失：ListReleases 每页 1→30 条，PollReleases 遍历所有比游标新的 release 逐个事件化，单轮补发上限 5 条，达上限或落库失败不推进游标、下轮续补（事件指纹幂等，重复扫描不重复投递）
- release 轮询翻页补拉：ListReleases 支持分页（30 条/页、ETag 仅第 1 页），PollReleases 逐页翻到游标所在页，中断期间超过一页的 release 也不再丢失旧版本
- release ID 溢出防护：events.subject_number int→int64（Ent schema + 双轨迁移，PG 侧 ALTER TYPE bigint，SQLite 无 DDL），GitHub release ID 不再受 PG int4 上限约束
- 前端最近事件面板：release 事件无 repository_id 时回退 payload_summary.repository 展示仓库名，测试覆盖回退与去重场景

+++ internal/normalizer/process.go
@@ -545,7 +545,7 @@ func (p *Processor) processWorkItem(ctx context.Context, repo store.Repository, 
 		p.eventDuplicate(repo, kind, normalizeAction(action))
 		return Result{Repository: &repo, Updated: false, SuppressNotify: suppress}, nil
 	}
-	num := saved.Number
+	num := int64(saved.Number)
 	srcUpdated := saved.SourceUpdatedAt
 	ev := store.Event{
 		ID: ulid.Make().String(), Source: "webhook", Kind: kind, Action: normalizeAction(action),
@@ -767,7 +767,7 @@ func (p *Processor) processSecurityAlert(ctx context.Context, kind string, env e
 		p.eventDuplicate(repo, kind, normalizeAction(env.Action))
 		return Result{Repository: &repo, Updated: false, SuppressNotify: suppress}, nil
 	}
-	num := a.Number
+	num := int64(a.Number)
 	srcUpdated := updatedAt
 	ev := store.Event{
 		ID: ulid.Make().String(), Source: "webhook", Kind: kind, Action: normalizeAction(env.Action),
